+++ hyprxa/dependencies/util.py
@@ -33,6 +33,8 @@ def parse_timestamp(query: Query, default_timedelta: int | None = None):
     else:
         default_timedelta = timedelta(seconds=0)
     def wrapper(time: str | None = query, timezone: str = DEFAULT_TIMEZONE) -> datetime:
+        timezone = timezone or DEFAULT_TIMEZONE
+        
         now = pendulum.now(tz="UTC").replace(tzinfo=None)
         if not time:
             return now - default_timedelta